chore(e2e): integration test example
