--- src/Libs/Features.Debugging/DebugRootFeature.cs
@@ -1,11 +1,0 @@
-﻿// using Systems.Debugging;
-//
-// namespace Features.Debugging;
-//
-// public sealed class DebugRootFeature : Entitas.Extended.Feature
-// {
-//     public DebugRootFeature(DrawRectangleCollisionComponentsSystem drawRectangleCollisionComponentsSystem)
-//     {
-//         Add(drawRectangleCollisionComponentsSystem);
-//     }
-// }
