--- src/Libs/Features/CameraFeature.cs
@@ -1,11 +1,0 @@
-﻿// using Entitas.Extended;
-//
-// namespace Features;
-//
-// public sealed class CameraFeature : Entitas.Extended.Feature
-// {
-//     public CameraFeature(IDrawSystem drawSystem)
-//     {
-//         Add(drawSystem);
-//     }
-// }
