--- src/Libs/Features/InputFeature.cs
@@ -1,11 +1,0 @@
-﻿// using Systems;
-//
-// namespace Features;
-//
-// public sealed class InputFeature : Entitas.Extended.Feature
-// {
-//     public InputFeature(InputSystem inputSystem)
-//     {
-//         Add(inputSystem);
-//     }
-// }
