+++ src/Libs/Features/RenderFeature.cs
@@ -1,0 +1,14 @@
+﻿using Scellecs.Morpeh;
+using Scellecs.Morpeh.Extended;
+using Systems.Render;
+
+namespace Features;
+
+public class RenderFeature : Feature
+{
+    public RenderFeature(World world,
+        RenderCharacterMovementAnimationSystem renderCharacterMovementAnimationSystem) : base(world)
+    {
+        Add(renderCharacterMovementAnimationSystem);
+    }
+}
